Java Asset Management text based app
This app was created in netbeans IDE and is a fully functioning end to
end stock management system. This system connects to a mySQl db so a
program like Xamp is needed to run the application. This is app also
handles multiple different file io read and write types for reporting.
This is text based only and contains no UI at present. It is a Java
version of the Asset Management website also on my profile.
https://github.com/seanjohn85/Asset-Management-WebSite